requirements.txt
Django==1.11.4
django-appconf==1.0.2
django-ckeditor==5.3.0
django-contact-form==1.4.2
django-countries==4.6.1
django-geojson==2.10.0
django-imagekit==4.0.2
django-js-asset==0.1.1
django-leaflet==0.22.0
django-maintenancemode==0.11.2
django-material==1.1.1
django-registration==2.2
IPy==0.83
jsonfield==2.0.2
olefile==0.44
pilkit==2.0
Pillow==4.2.1
pkg-resources==0.0.0
python-magic==0.4.13
pytz==2017.2
six==1.10.0
